+++ source/app/service-providers/config/get-config-template.ts
@@ -137,6 +137,7 @@ export interface ConfigOptions {
     inputMode: 'default'|'vim'|'emacs'
     boldFormatting: '**'|'__'
     italicFormatting: '_'|'*'
+    highlightFormatting: 'span'|'=='
     readabilityAlgorithm: 'dale-chall'|'gunning-fog'|'coleman-liau'|'automated-readability'
     lint: {
       markdown: boolean
@@ -171,6 +172,7 @@ export interface ConfigOptions {
     theme: MarkdownTheme
     hideToolbarInDistractionFree: boolean
     markdownFileExtensions: boolean
+    previewModeShowSyntaxWhenCursorIsAdjacent: boolean
     imageWidth: number
     imageHeight: number
     renderingMode: 'preview'|'raw'
@@ -197,6 +199,7 @@ export interface ConfigOptions {
     msoffice: FileTypeSettings<boolean, boolean, 'system'>
     openOffice: FileTypeSettings<boolean, boolean, 'system'>
     dataFiles: FileTypeSettings<boolean, boolean, 'system'>
+    dotFiles: FileTypeSettings<boolean, boolean>
   }
   watchdog: {
     activatePolling: boolean
@@ -226,6 +229,7 @@ export interface ConfigOptions {
     showNewFileButton: boolean
     showPreviousFileButton: boolean
     showNextFileButton: boolean
+    showPandocDivSpanButton: boolean
     showMarkdownCommentButton: boolean
     showMarkdownLinkButton: boolean
     showMarkdownImageButton: boolean
@@ -341,6 +345,7 @@ export function getConfigTemplate (): ConfigOptions {
       inputMode: 'default', // Can be default, vim, emacs
       boldFormatting: '**', // Can be ** or __
       italicFormatting: '_', // Can be * or _
+      highlightFormatting: '==', // Can be 'span' or ==
       readabilityAlgorithm: 'dale-chall', // The algorithm to use with readability mode.
       showStatusbar: true,
       showFormattingToolbar: true,
@@ -437,6 +442,7 @@ export function getConfigTemplate (): ConfigOptions {
       theme: 'berlin', // The theme, can be berlin|frankfurt|bielefeld|karl-marx-stadt|bordeaux
       hideToolbarInDistractionFree: false,
       markdownFileExtensions: false,
+      previewModeShowSyntaxWhenCursorIsAdjacent: true,
       imageWidth: 100, // Maximum preview image width
       imageHeight: 50, // Maximum preview image height
       renderingMode: 'preview',
@@ -457,7 +463,8 @@ export function getConfigTemplate (): ConfigOptions {
       pdf: { showInFilemanager: false, showInSidebar: true, openWith: 'system' },
       msoffice: { showInFilemanager: false, showInSidebar: true, openWith: 'system' },
       openOffice: { showInFilemanager: false, showInSidebar: true, openWith: 'system' },
-      dataFiles: { showInFilemanager: false, showInSidebar: true, openWith: 'system' }
+      dataFiles: { showInFilemanager: false, showInSidebar: true, openWith: 'system' },
+      dotFiles: { showInFilemanager: false, showInSidebar: false, openWith: 'system' }
     },
     // Language
     selectedDicts: [], // By default no spell checking is active to speed up first start.
@@ -481,6 +488,7 @@ export function getConfigTemplate (): ConfigOptions {
       showNewFileButton: true,
       showPreviousFileButton: true,
       showNextFileButton: true,
+      showPandocDivSpanButton: true,
       showMarkdownCommentButton: true,
       showMarkdownLinkButton: true,
       showMarkdownImageButton: true,