+++ source/common/modules/markdown-editor/plugins/highlight-whitespace.ts
@@ -30,7 +30,7 @@ import {
   WidgetType
 } from '@codemirror/view'
 import { configUpdateEffect } from '../util/configuration'
-import { syntaxTree } from '@codemirror/language'
+import { foldedRanges, syntaxTree } from '@codemirror/language'
 
 const extensionCompartment = new Compartment()
 
@@ -42,6 +42,7 @@ class PilcrowWidget extends WidgetType {
     const span = document.createElement('span')
     span.className = 'cm-pilcrow'
     span.textContent = '¶'
+
     span.setAttribute('aria-hidden', 'true')
 
     return span
@@ -55,16 +56,42 @@ class PilcrowWidget extends WidgetType {
 // The widget should appear last, so we set it to the highest side
 const pilcrowDeco = Decoration.widget({ widget: new PilcrowWidget(), side: 10000 })
 
+const EXCLUDED_NODES = [
+  'CodeText',
+  'CodeBlock',
+  'FencedCode',
+  'YAMLFrontmatter',
+  'PandocDivMark',
+  'PandocDivInfo',
+  'PandocAttribute'
+]
+
 function showLineEndings (view: EditorView): DecorationSet {
   const builder = new RangeSetBuilder<Decoration>()
   for (const { from, to } of view.visibleRanges) {
     for (let pos = from; pos <= to;) {
       const line = view.state.doc.lineAt(pos)
-      const node = syntaxTree(view.state).resolve(line.to, -1)
-      if (![ 'CodeText', 'CodeBlock', 'FencedCode' ].includes(node.name)) {
-        builder.add(line.to, line.to, pilcrowDeco)
-      }
       pos = line.to + 1
+
+      // Do not draw pilcrows on folded lines so that, when folded,
+      // two pilcrows do not appear on the same line.
+      let isFolded = false
+      foldedRanges(view.state).between(line.from, line.to, (from, to) => {
+        if (pos >= from && pos <= to) {
+          isFolded = true
+          return false
+        }
+      })
+
+      if (isFolded) { continue }
+
+      const node = syntaxTree(view.state).resolve(line.to, -1)
+
+      if (EXCLUDED_NODES.includes(node.name) || (node.parent && EXCLUDED_NODES.includes(node.parent.name))) {
+        continue
+      }
+
+      builder.add(line.to, line.to, pilcrowDeco)
     }
   }
   return builder.finish()
@@ -91,8 +118,22 @@ const pilcrowPlugin = ViewPlugin.fromClass(class {
 
 const pilcrowTheme = EditorView.baseTheme({
   '.cm-pilcrow': {
+    // By setting `display: inline-block` and
+    // `width: 0px`, we prevent the pilcrow
+    // from triggering line wrapping
+    display: 'inline-block',
+    width: '0px',
+    // Overrides styling from the `visual-indent` plugin
+    // which otherwise causes the pilcrow to be offset into
+    // the text.
+    textIndent: 'initial',
+    // These help prevent issues related to
+    // pilcrows inserting literal text content
+    userSelect: 'none',
+    pointerEvents: 'none',
+    // Styling
     color: '#aaa',
-    opacity: '0.5'
+    opacity: '0.5',
   }
 })