+++ source/common/modules/markdown-editor/renderers/render-emphasis.ts
@@ -18,6 +18,7 @@ import { type RangeSet, type Range } from '@codemirror/state'
 import { type ViewUpdate, type EditorView, ViewPlugin, Decoration, type DecorationSet, WidgetType } from '@codemirror/view'
 import { rangeInSelection } from '../util/range-in-selection'
 import type { SyntaxNode } from '@lezer/common'
+import { configField } from '../util/configuration'
 
 class BulletWidget extends WidgetType {
   constructor (readonly node: SyntaxNode) {
@@ -41,7 +42,7 @@ export class SpaceWidget extends WidgetType {
     super()
   }
 
-  eq (other: BulletWidget): boolean {
+  eq (other: SpaceWidget): boolean {
     if (this.node === undefined || other.node === undefined) {
       return false
     }
@@ -52,6 +53,7 @@ export class SpaceWidget extends WidgetType {
   toDOM (_view: EditorView): HTMLElement {
     const elem = document.createElement('span')
     elem.innerHTML = '&nbsp;'.repeat(this.numChars)
+    elem.classList.add('rendered-space')
     return elem
   }
 }
@@ -59,6 +61,7 @@ export class SpaceWidget extends WidgetType {
 function hideFormattingCharacters (view: EditorView): RangeSet<Decoration> {
   const ranges: Array<Range<Decoration>> = []
   const hiddenDeco = Decoration.replace({})
+  const includeAdjacent = view.state.field(configField, false)?.previewModeShowSyntaxWhenCursorIsAdjacent ?? true
 
   for (const { from, to } of view.visibleRanges) {
     syntaxTree(view.state).iterate({
@@ -66,7 +69,7 @@ function hideFormattingCharacters (view: EditorView): RangeSet<Decoration> {
       to,
       enter (node) {
         // Do not hide any characters if a selection is inside here
-        if (rangeInSelection(view.state.selection, node.from, node.to, true)) {
+        if (rangeInSelection(view.state.selection, node.from, node.to, includeAdjacent)) {
           return
         }
 
@@ -146,8 +149,12 @@ function hideFormattingCharacters (view: EditorView): RangeSet<Decoration> {
             }
 
             // Only render QuoteMark if the parent does not contain a cursor.
-            if (parentNode && !rangeInSelection(view.state.selection, parentNode.from, parentNode.to, true)) {
-              ranges.push(Decoration.replace({ widget: new SpaceWidget(node.to - node.from, node.node) }).range(node.from, node.to))
+            if (parentNode && !rangeInSelection(view.state.selection, parentNode.from, parentNode.to, includeAdjacent)) {
+              // We want to also hide any trailing whitespace. Since quotemarks
+              // can be followed by a max of 3 spaces, we grab three more characters
+              // and test the resulting string
+              const match = /^(\>[ ]{0,3})/.exec(view.state.sliceDoc(node.from, node.from + 3))
+              ranges.push(hiddenDeco.range(node.from, match ? node.from + match[1].length : node.to ))
             }
 
             break